add create folder Reports

+++ helper/exportToJsonFile.ts
@@ -1,5 +1,10 @@
 export const toJsonFile = (jsonData: any, name: string = "") => {
   const fs = require("fs");
   let data = JSON.stringify(jsonData);
-  fs.writeFileSync(`reports/performanceReport${name}.json`, data);
+  if (fs.existsSync(`reports`)) {
+    fs.writeFileSync(`reports/performanceReport${name}.json`, data);
+  } else {
+    fs.mkdirSync("reports");
+    fs.writeFileSync(`reports/performanceReport${name}.json`, data);
+  }
 };
